Fixed editor compile
