+++ src/cpp/editor/gui_editor.h
@@ -1,0 +1,10 @@
+#ifndef GUI_EDITOR_H
+#define GUI_EDITOR_H
+
+#include "gui/gui_utils.h"
+
+void __fastcall NormalDialog(char *msg, int a2, int x, int y, int img1Type, int img1Arg, int img2Type, int img2Arg, int writeOr, signed int a10);
+
+#pragma pack(pop)
+
+#endif
